rename a few files
