+++ src/fontra/client/core/glyph-svg.js
